% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomDiffError.R
\name{dominant.diff}
\alias{dominant.diff}
\title{This function returns the non-centrality parameter when the genotyping error is different in cases
and in controls, assuming dominant mode of inheritance.}
\usage{
dominant.diff(r, pA, ncase, nctrl, pD, error.mat.case, error.mat.ctrl)
}
\arguments{
\item{r}{relative risk}

\item{pA}{minor allele frequency}

\item{ncase}{number of cases}

\item{nctrl}{number of control samples}

\item{pD}{disease prevalence}

\item{error.mat.case}{3*3 matrix of genotyping error in cases}

\item{error.mat.ctrl}{3*3 matrix of genotyping error in controls}
}
\value{
non-centrality parameter of the chi-square distribution under alternative hypothesis
}
\description{
This function returns the non-centrality parameter when the genotyping error is different in cases
and in controls, assuming dominant mode of inheritance.
}

