% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiDiffError.R
\name{NCP.diff.null}
\alias{NCP.diff.null}
\title{This function calculates the non-centrality parameter of burden test statistic under null hypothesis
when the genotyping error is different in cases and controls.}
\usage{
NCP.diff.null(pA, pD, m10.case, m00.case, m10.ctrl, m00.ctrl, M, Ncase)
}
\arguments{
\item{pA}{minor allele frequency}

\item{pD}{disease prevalence}

\item{m10.case}{1-sensitivity in cases}

\item{m00.case}{specificity in cases}

\item{m10.ctrl}{1-sensitivity in controls}

\item{m00.ctrl}{specificity in controls}

\item{M}{number of markers in a gene/region}

\item{Ncase}{number of cases}
}
\value{
the non-centrality parameter of burden test statistic under null hypothesis
}
\description{
This function calculates the non-centrality parameter of burden test statistic under null hypothesis
when the genotyping error is different in cases and controls.
}

