% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCP.R
\name{NCP.single}
\alias{NCP.single}
\title{This function calculates non-centrality parameter for single marker analysis.}
\usage{
NCP.single(r, pA, ncase, nctrl, pD, emat1, emat2 = emat1, type = "ndiff",
  moi = "m")
}
\arguments{
\item{r}{relative risk}

\item{pA}{minor allele frequency}

\item{ncase}{number of cases}

\item{nctrl}{number of controls}

\item{pD}{disease prevalence}

\item{type}{"ndiff"/"diff", indicating non-differential or differential genotyping error}

\item{moi}{"m"/"r"/"d", indicating multiplicative, or recessive, or dominant mode of inheritance}

\item{error.mat1}{3*3 matrix of genotyping error in cases}

\item{error.mat2}{3*3 matrix of genotyping error in controls}
}
\value{
the non-centrality parameter of the chisquare distribution
}
\description{
This function calculates non-centrality parameter for single marker analysis.
}

