% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiNoDiffError.R
\name{PhiCtrlWES}
\alias{PhiCtrlWES}
\title{This function calculates the proportion of subjects with at least one rare allele in controls
when the genotyping error is the same in cases and controls.}
\usage{
PhiCtrlWES(r, pA, pD, m10, m00, M)
}
\arguments{
\item{pA}{minor allele frequency}

\item{pD}{disease prevalence}

\item{m10}{1-sensitivity}

\item{m00}{specificity}

\item{M}{number of markers in a gene/region}
}
\value{
the proportion of subjects with at least one rare allele in cases
}
\description{
This function calculates the proportion of subjects with at least one rare allele in controls
when the genotyping error is the same in cases and controls.
}

