% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{power.gwas}
\alias{power.gwas}
\title{This function returns the power of chi-square test given type I error and non-centrality parameter}
\usage{
power.gwas(alpha, alt.ncp, null.ncp = 0)
}
\arguments{
\item{alpha}{level of type I error}

\item{alt.ncp}{non-centrality parameter under alternative hypothesis}

\item{null.ncp}{non-centrality parameter under null hypohtesis}
}
\value{
power of chi-square test
}
\description{
This function returns the power of chi-square test given type I error and non-centrality parameter
}

