% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{Power.multi}
\alias{Power.multi}
\title{This function calculates the statistical power for multiple marker analysis.}
\usage{
Power.multi(alpha, r, pA, ncase, pD, M, m10.case, m00.case,
  m10.ctrl = m10.case, m00.ctrl = m00.case, type = "ndiff")
}
\arguments{
\item{alpha}{type I error}

\item{r}{relative risk}

\item{pA}{minor allele frequency}

\item{ncase}{number of cases (assume number of controls is the same as cases)}

\item{pD}{disease prevalence}

\item{M}{number of markers in a gene/region}

\item{m10.case}{1-sensitivity in cases}

\item{m00.case}{specificity in cases}

\item{m10.ctrl}{1-sensitivity in controls}

\item{m00.ctrl}{specificity in controls}

\item{type}{"ndiff"/"diff", indicating non-differential or differential genotyping error}
}
\value{
the non-centrality parameter of the chisquare distribution
}
\description{
This function calculates the statistical power for multiple marker analysis.
}

