% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Power.R
\name{Power.single}
\alias{Power.single}
\title{This function calculates the statistical power for single marker analysis.}
\usage{
Power.single(alpha, r, pA, ncase, nctrl, pD, emat1, emat2 = emat1,
  type = "ndiff", moi = "m")
}
\arguments{
\item{alpha}{type I error}

\item{r}{relative risk}

\item{pA}{minor allele frequency}

\item{ncase}{number of cases}

\item{nctrl}{number of controls}

\item{pD}{disease prevalence}

\item{type}{"ndiff"/"diff", indicating non-differential or differential genotyping error}

\item{moi}{"m"/"r"/"d", indicating multiplicative, or recessive, or dominant mode of inheritance}

\item{error.mat1}{3*3 matrix of genotyping error in cases}

\item{error.mat2}{3*3 matrix of genotyping error in controls}
}
\value{
the statistical power of single marker association test
}
\description{
This function calculates the statistical power for single marker analysis.
}

