% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecessiveNoDiffError.R
\name{recessive.ndiff}
\alias{recessive.ndiff}
\title{This function calculates the non-centrality parameter of the chisquare distribution
when the genotyping error is the same in cases and controls, 
assuming recessive mode of inheritance}
\usage{
recessive.ndiff(r, pA, ncase, nctrl, pD, error.mat)
}
\arguments{
\item{r}{relative risk}

\item{pA}{minor allele frequency}

\item{ncase}{number of cases}

\item{nctrl}{number of controls}

\item{pD}{disease prevalence}

\item{error.mat}{3*3 matrix of genotyping error}
}
\value{
the non-centrality parameter of the chisquare distribution

assuming the genotyping error is the same in cases and controls
}
\description{
This function calculates the non-centrality parameter of the chisquare distribution
when the genotyping error is the same in cases and controls, 
assuming recessive mode of inheritance
}

