% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SingleMarker.R
\name{SingleMarker}
\alias{SingleMarker}
\title{This function is called in burden test}
\usage{
SingleMarker(r, pA, pD)
}
\arguments{
\item{r}{relative risk}

\item{pA}{allele frequency}

\item{pD}{disease prevalence}
}
\value{
genotype distribution break down by cases and controls
}
\description{
This function is called in burden test
}

