% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinSize.R
\name{Size.multi}
\alias{Size.multi}
\title{This function calculates the minimum sample size needed to achieve a specified level 
of power for multiple marker analysis.}
\usage{
Size.multi(alpha, p, r, pA, pD, M, m10.case, m00.case, m10.ctrl = m10.case,
  m00.ctrl = m00.case, type = "ndiff", range = c(1, 1e+05))
}
\arguments{
\item{alpha}{type I error}

\item{p}{pre-specified statistical power}

\item{r}{relative risk}

\item{pA}{minor allele frequency}

\item{pD}{disease prevalence}

\item{M}{number of markers in a gene/region}

\item{m10.case}{1-sensitivity in cases}

\item{m00.case}{specificity in cases}

\item{m10.ctrl}{1-sensitivity in controls}

\item{m00.ctrl}{specificity in controls}

\item{type}{"ndiff"/"diff", indicating non-differential or differential genotyping error}
}
\value{
the minimum sample size needed to achieve a specified level

of power for multiple marker analysis.
}
\description{
This function calculates the minimum sample size needed to achieve a specified level 
of power for multiple marker analysis.
}

